Brainrot/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Cython==3.0.11
decorator==4.4.2
distro==1.9.0
dtw-python==1.5.3
ffmpeg-python==0.2.0
filelock==3.16.1
fsspec==2024.10.0
future==1.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
jiter==0.7.0
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai==1.54.3
openai-whisper==20240930
packaging==24.1
pillow==11.0.0
proglog==0.1.10
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
scipy==1.14.1
setuptools==75.3.0
shellingham==1.5.4
sniffio==1.3.1
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
tqdm==4.67.0
typer==0.13.0
typing_extensions==4.12.2
urllib3==2.2.3
websockets==12.0
whisper-timestamped==1.15.6